+++ Tezo.Todo.Data/Program.cs
@@ -1,0 +1,33 @@
+using Microsoft.EntityFrameworkCore;
+using Tezo_Todo.Data;
+var builder = WebApplication.CreateBuilder(args);
+
+// Add services to the container.
+
+builder.Services.AddControllers();
+// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
+//builder.Services.AddDbContext<TodoAPIDbContext>(options => options.UseInMemoryDatabase("TODO"));
+
+builder.Services.AddDbContext<TodoAPIDbContext>(options =>
+               options.UseNpgsql(builder.Configuration.GetConnectionString("TodoApiConnectionString")));
+
+builder.Services.AddAutoMapper(AppDomain.CurrentDomain.GetAssemblies());
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
+
+app.UseHttpsRedirection();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
