+++ Tezo.Todo.Models/Models/Assignment.cs
@@ -1,0 +1,27 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace Tezo_Todo.Models
+{
+    public class Assignment
+    {
+        [Key]
+        public Guid Id { get; set; }
+        public string Title { get; set; }
+        public string Description { get; set; }
+        public DateTime DueDate { get; set; }
+        public Status Status { get; set;  }
+        public Priority Priority { get; set; }
+
+        [ForeignKey("User")]
+        public Guid UserId { get; set; }
+        public virtual User User { get; set; }
+
+
+        public Assignment()
+        {
+
+        }
+
+    }   
+}
