+++ Tezo.Todo.Models/Models/enum.cs
@@ -1,0 +1,16 @@
+﻿namespace Tezo_Todo.Models
+{
+    public enum Status
+    {
+        NotStarted,
+        InProgress,
+        Completed
+    }
+
+    public enum Priority
+    {
+        Low,
+        Medium,
+        High
+    }
+}
