+++ Tezo_Todo/Models/PaginatedList.cs
@@ -1,19 +1,19 @@
-﻿namespace Tezo_Todo.Models
-{
-    public class PaginatedList<T>
-    {
-        // Model for Pagination
-        public List<T> Items { get; }
-        public int PageIndex { get; }
-        public int TotalPages { get; }
-        public bool HasPreviousPage => PageIndex > 1;
-        public bool HasNextPage => PageIndex < TotalPages;
+﻿//namespace Tezo_Todo.Models
+//{
+//    public class PaginatedList<T>
+//    {
+//        // Model for Pagination
+//        public List<T> Items { get; }
+//        public int PageIndex { get; }
+//        public int TotalPages { get; }
+//        public bool HasPreviousPage => PageIndex > 1;
+//        public bool HasNextPage => PageIndex < TotalPages;
 
-        public PaginatedList(List<T> items, int pageIndex, int totalPages)
-        {
-            Items = items;
-            PageIndex = pageIndex;
-            TotalPages = totalPages;
-        }
-    }
-}
+//        public PaginatedList(List<T> items, int pageIndex, int totalPages)
+//        {
+//            Items = items;
+//            PageIndex = pageIndex;
+//            TotalPages = totalPages;
+//        }
+//    }
+//}
