+++ Tezo_Todo/Models/enum.cs
@@ -1,16 +1,16 @@
-﻿namespace Tezo_Todo.Models
-{
-    public enum Status
-    {
-        NotStarted,
-        InProgress,
-        Completed
-    }
+﻿//namespace Tezo_Todo.Models
+//{
+//    public enum Status
+//    {
+//        NotStarted,
+//        InProgress,
+//        Completed
+//    }
 
-    public enum Priority
-    {
-        Low,
-        Medium,
-        High
-    }
-}
+//    public enum Priority
+//    {
+//        Low,
+//        Medium,
+//        High
+//    }
+//}
