+++ Tezo_Todo/Services/Interfaces/IUserServices.cs
@@ -1,13 +1,13 @@
-﻿using Tezo_Todo.Models;
+﻿//using Tezo_Todo.Models;
 
-namespace Tezo_Todo.Services.Interfaces
-{
-    public interface IUserServices
-    {
-        public List<User> GetAllUser();
-        public User AddUser(User user);
-        public Task<User> GetUserById(Guid id);
-        public Task<bool> UpdateUser(Guid id, User user);
-        public Task<User> DeleteUser(Guid id);
-    }
-}
+//namespace Tezo_Todo.Services.Interfaces
+//{
+//    public interface IUserServices
+//    {
+//        public List<User> GetAllUser();
+//        public User AddUser(User user);
+//        public Task<User> GetUserById(Guid id);
+//        public Task<bool> UpdateUser(Guid id, User user);
+//        public Task<User> DeleteUser(Guid id);
+//    }
+//}
