+++ Tezo_Todo/Tezo.Todo.csproj
@@ -23,4 +23,10 @@
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Tezo.Todo.Data\Tezo.Todo.Data.csproj" />
+    <ProjectReference Include="..\Tezo.Todo.Models\Tezo.Todo.Models.csproj" />
+    <ProjectReference Include="..\Tezo.Todo.Services\Tezo.Todo.Services.csproj" />
+  </ItemGroup>
+
 </Project>
